Lab 3 Task 2 Done
Page layout with flexbox finished.

In order to display both variants of the same page I had to create the
second .html file. That's why there is this much new files.

It is a bad practice and it would be better to use JavaScript and
button to change used .css files for the same .html file.
This will be done later.